Use auto-generated constants instead of globals.
This is a large change that touches all files in the factory, but at the
same time it is mostly mechanical and doesn't affect the logic much.

The primary change is removing the "load X" scripts that used to set
global variables, and replacing them with two new scripts:
"factory macros" and "factory constants". These scripts are compile-time
only, and not part of the bundle.

"factory constants" lists all the craftable items, in order. It takes
over from the list in "launch factory craft" as the canonical list, so
the README is updated accordingly. "factory macros" use Lua macros to
automagically create :const values for each item. By using constants
instead of globals, constant-folding kicks in and we save a lot of
space: From 309kB down to 256kB!

Some uses of i2s() have also been removed, to get constant-folding
to apply to those expressions. Generally it's ones where the entire
expression can be folded down to a single string constant.

Using "factory constants" everywhere depends on the :import mechanism
which is only supported by my enhanced-syntax editor. (The same goes for
Lua-based macros.) IMO it's a worthwhile upgrade, especially since there
are further improvements that can be built off of this.