Use a more efficient variable-name encoding.
This shortens the name-prefix from "craft_queue_" down to "cq" and also
changes the encoding formula to be item * 10 + tier instead of
tier * 1000 + item. This allows for more opportunities for
constant-folding, as well as being better suited for the auto-generated
loops that are more and more part of the package.

This also includes many minor tweaks to recipes to improve folding and
reduce arithmetic nodes.

All told, reduces bundle import size to 172Kb. The import is now
noticably faster than when I started.