Replace recipe scripts with Lua-generated recipe system
This is a major change that touches all the recipe-involved code, which
is most of what remains. (Although producer-related code is mostly
untouched.) Instead of having hardcoded recipes, all the recipes are
declared in a psuedo-language made of macro calls, in a large block in
the new file "recipes".

This file is also responsible for actually evaluating the recipes. In
addition, it absorbs the code for counting items: This all becomes part
of a single table-driven loop, that runs over recipe data, counting
existing items and summing queue entries as needed.

This drops the import size down to 58Kb, around a third of what it was
in the previous commit and less than a fifth from when I started. It
imports *noticably* faster, and the text actually shows up in the box
after pasting now. The total number of imported scripts is also down to 9.

There is a *lot* of new Lua code involved to make this happen. Some of
it might need more comments; it is unfortunately awkward to put comments
inline in the code due to how macros force backslash-based line
continuation, but it's not impossible.