Bugfix a subtle race condition
When not immediately re-registering, turbo_register was jumping to the
preamble where it sets the initial values. This was mostly benign, but
there was a race condition where the effect of `turbo start` could have
been overwritten by the write of "turbo.register = 0".

I think this was just an oversight, since the jump target was meant to
be the waituntil.