Update turbo exec to V2.2
All names are updated to be part of packages. A compatibility shim is
kept with a copy of "turbo start" and "turbo stop" being left outside
the package, so that V2.1 scripts will continue to work.

The docs are *not* updated in this commit, it's just the code. Code has
been tested, and appears to work correctly both with and without a copy
of V2.1 installed.