Optimize the hot parts of turbo exec
This uses my findings from benchmarking the AI, namely that (for
functions that don't interact with the game state) performance is
strictly determined by number of nodes, and there is no short-circuiting
in conditionals.

As a result, finding ways to avoid the trinary comparison, boolean, and
arithmetic functions is best, when possible (and where it makes a
difference).